requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
absl-py==2.3.1
    # via mediapipe
attrs==25.4.0
    # via mediapipe
cffi==2.0.0
    # via sounddevice
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.20.2
    # via torch
flatbuffers==25.12.19
    # via mediapipe
fonttools==4.61.1
    # via matplotlib
fsspec==2025.12.0
    # via torch
jax==0.8.2
    # via mediapipe
jaxlib==0.8.2
    # via
    #   jax
    #   mediapipe
jinja2==3.1.6
    # via torch
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via
    #   hand-expression (pyproject.toml)
    #   mediapipe
mediapipe==0.10.13
    # via hand-expression (pyproject.toml)
ml-dtypes==0.5.4
    # via
    #   jax
    #   jaxlib
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.2.6
    # via
    #   contourpy
    #   jax
    #   jaxlib
    #   matplotlib
    #   mediapipe
    #   ml-dtypes
    #   opencv-contrib-python
    #   opencv-python
    #   scipy
    #   torchvision
opencv-contrib-python==4.12.0.88
    # via mediapipe
opencv-python==4.12.0.88
    # via hand-expression (pyproject.toml)
opt-einsum==3.4.0
    # via jax
packaging==25.0
    # via matplotlib
pillow==12.1.0
    # via
    #   matplotlib
    #   torchvision
protobuf==4.25.8
    # via mediapipe
pycparser==2.23
    # via cffi
pyparsing==3.3.1
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
scipy==1.16.3
    # via
    #   jax
    #   jaxlib
setuptools==80.9.0
    # via torch
six==1.17.0
    # via python-dateutil
sounddevice==0.5.3
    # via mediapipe
sympy==1.14.0
    # via torch
torch==2.9.1
    # via
    #   hand-expression (pyproject.toml)
    #   torchvision
torchvision==0.24.1
    # via hand-expression (pyproject.toml)
tqdm==4.67.1
    # via hand-expression (pyproject.toml)
typing-extensions==4.15.0
    # via torch
